feat(core/presentation): Add helper functions for generating categorized validation messages
const v = new FormValidator(values);
v.field('foo').withValidators(value => {
  return value === 'foo' && errorMessage('NO, NOT FOO!!!')
});

also moved all category stuff into categories.ts
