+++ app/scripts/modules/core/src/presentation/forms/validation/index.ts
@@ -1,3 +1,4 @@
+export * from './categories';
 export * from './validators';
 export * from './validation';
 export * from './FormValidator';
